refactor: thing: Separate controller from handler
In order to improve the code readability, the code responsible to
handle requests that operates on the things was moved to a specific
controller package. After this change, both thing and user packages
have the same structure, which represent the correct architectual
concepts.

As a result of, the code related to the server side messaging lifecycle
and subscriptions to the topics (exchange - binding key) was moved to
the server package.
